Change versioning of composer files
This commit also moves `version::Error` enum into a separate module `version::error::Error` to make it easier to make versioning modules that can return errors.
